backend/requirements.txt
Django>=5.1,<5.2
psycopg[binary]>=3.1,<3.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.3,<24.4
python-dotenv>=1.0
